pyproject.toml
[tool.poetry]
name = "stackexchange_django"
version = "0.1.0"
description = ""
authors = ["Kostas Kokkoros"]

[tool.poetry.dependencies]
python = "^3.10"
celery = "*"
Django = "~4.0" # Django extensions not compatible with 4.1, https://github.com/django-extensions/django-extensions/issues/1756 must be released first
redis = "*"
defusedxml = "*"
django-celery-results = "*"
django-environ = "*"
djangorestframework = "*"
drf-spectacular = "*"
psycopg2 = "*"
py7zr = "*"
requests = "*"
tqdm = "*"

[tool.poetry.dev-dependencies]
bandit = "*"
coverage = "*"
django-debug-toolbar = "*"
django-extensions = "*"
factory-boy = "*"
pylint = "*"
pylint-django = "*"
safety = "*"
Werkzeug = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
